Currently when we get a plan deletion request from RPSaaS, we always accept it regardless of if the plan actually exists in our DB or not.  This causes an issue when deleting PPE and Dev plans from the Azure portal, because the portal always hit's Prod.    If this mismatched delete happens, we aren't able to delete the plan even from the VSO extension or portal because ARM will see that the resource doesn't exist and not send VSO another delete request.  The effect of that is that ARM will delete the plan, but it won't be deleted on the VSO side which causes issues with billing and also means plans aren't properly cleaned up.

This change will cause the FE to validate that the plan actually exists in our DB before approving it to be deleted.  Soft deleted plans are included in the lookup as well to cover the case where a delete request partially succeeds and VSO deletes the plan, but ARM does not.  If the FE receives a delete request for a plan which does not exist in our DB at all, it will return an error response and ARM will not delete the plan.

I tested the portal behavior for these cases in the Dogfood portal (pointing Dogfood prod endpoint to localhost):

- Plan exists in DB, not soft-deleted => Delete succeeds
- Plan exists in DB, is soft-deleted => Delete succeeds
- Plan does not exist in DB => Delete fails, error appears pointing to Activity Log which shows no info (see gif)

![deletePlan.gif](https://dev.azure.com/devdiv/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/240604/attachments/deletePlan.gif)

+++ src/FrontEnd/FrontEndWebApi/Src/FrontEndWebApi/Controllers/SubscriptionsController.cs
@@ -282,22 +282,29 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.FrontEndWebApi.Controllers
                     ValidationUtil.IsTrue(ResourceTypeIsValid(resourceType));
                     ValidationUtil.IsTrue(ResourceProviderIsValid(providerNamespace));
 
-                    var plan = new VsoPlanInfo
+                    var planInfo = new VsoPlanInfo
                     {
                         Name = resourceName,
                         ResourceGroup = resourceGroup,
                         Subscription = subscriptionId,
                     };
 
+                    var plan = await planManager.GetAsync(planInfo, logger, includeDeleted: true);
+
+                    if (plan == null)
+                    {
+                        return CreateErrorResponse("ResourceDeleteFailed", "Plan not found");
+                    }
+
                     var environments = await environmentManager.ListAsync(
-                        logger, planId: plan.ResourceId);
+                        logger, planId: planInfo.ResourceId);
                     var nonDeletedEnvironments = environments.Where(t => t.State != CloudEnvironmentState.Deleted).ToList();
                     if (nonDeletedEnvironments.Any())
                     {
                         foreach (var environment in nonDeletedEnvironments)
                         {
                             var childLogger = logger.NewChildLogger()
-                                .AddVsoPlan(plan);
+                                .AddVsoPlan(planInfo);
 
                             _ = Task.Run(async () =>
                                 {
@@ -317,19 +324,20 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.FrontEndWebApi.Controllers
                                 });
                         }
 
-                        logger.AddVsoPlan(plan)
+                        logger.AddVsoPlan(planInfo)
                             .FluentAddValue("Count", $"{nonDeletedEnvironments.Count()}")
                             .LogInfo("plan_delete_environment_delete_success");
                     }
 
-                    var response = await planManager.DeleteAsync(plan, logger);
-                    if (!response)
+                    if (!plan.IsDeleted)
                     {
-                        logger.AddVsoPlan(plan).LogError("plan_delete_doesnotexist_error");
-                        return Ok();
+                        plan = await planManager.DeleteAsync(plan, logger);
+                        logger.AddVsoPlan(planInfo).LogInfo($"plan_delete_success");
                     }
-
-                    logger.LogInfo($"plan_delete_success");
+                    else
+                    {
+                        logger.AddVsoPlan(planInfo).LogWarning("plan_delete_alreadydeleted");
+                    }
 
                     // Required response format in case validation pass with empty body.
                     return Ok();
